requirements.txt
alembic==1.12.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coverage==7.3.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
gevent==23.9.1
greenlet==3.0.0rc3
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-mock==3.11.1
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.5
watchdog==3.0.0
Werkzeug==3.0.0
zerobouncesdk==1.1.0
zope.event==5.0
zope.interface==6.0
